Update name.js
refactoring of the name.js file. The utility of the function remains the same but to continue with testing this file needed to be divided into a main process with a supplementary function. As previously mentioned the utility remains the same and the file maintains the previous functionality of the previous iteration.