Create server.test.js
The server.test.js files are a singular test suite that contains three tests to ensure the proper function of the server file. The first test verifies the server is running on the correct port, the second test verifies that the entered name is being successfully set as the group's id, and the third test checks for an error to be displayed when an invalid input is entered.