Update deliverable_6_implementation2.md
Addition of Ben's requirements implementation, the tests to go along with in, the code quality section and two lessons added to the lesson's learned section. I removed a "!" from the link to Dalton's test files as it was trying to display an image where none was. Please edit anything that seems out of place or any typos present.